Do not close watchdog file descriptors when daemonizing.
When dsme is started with --daemon command line option, it does fork() and
then exit() from parent process and close() of all file descriptors in
child process. That results in watchdogs file descriptors to be closed and
watchdog resetting the device as watchdog is no longer opened.

Fix that by checking if we are going to close watchdog descriptor in child
process and not doing close() in that case.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>