lifeguard: Wait until process is really killed before announcing
Currently process exit response is send as soon as kill() is called. This
leads to various problems as by the time the requester gets response, the
process is actually still alive, so it's resources are not free - for
example dbus service names, or socket files. That prevents various
services (like mce) from being gracefully restarted.

Fix that by waiting for kill() to complete before sending a response.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>